+++ src/components/CodeHighlight/CHANGELOG.md
@@ -1,2 +1,3 @@
 - 2025-09-12: Initial version.
 - 2025-09-26: Add support for TSX language in CodeHighlight.
+- 2025-10-05: Add `withLineNumbers` prop to CodeHighlight component.